libdocument: use GtkFileChooser in add filters
The method to add filters to a GtkFileChooser,
`ev_document_factory_add_filters ()`, had the chooser parameter
specified as GtkWidget and cast it to GtkFileChooser where used. This
caused a runtime warning about an invalid cast.

To fix this the chooser parameter is specified as GtkFileChooser. The
calling methods create GtkFileChooserNative objects and these are
validly cast to GtkFileChooser on the call instead.

Fixes #1863.