shell: Port EvApplication to gtk4
* Switch to libadwaita
* Remove usage of GdkScreen
* Move several shortcuts to EvWindow

These shortcuts need to be move to EvWindow as a shortcut controller due
to the original event handling reorder hacks can't be implement in
GTK4. And we only move the shortcuts that conflicts with EvSearchBox.
For example, the shortcut `f` will be captured by EvWindow when you type
`f` with EvSearchBox opened.

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>