ev-stock-icons: remove GtkIconFactory code
This code comes from e90ec741051727680a653837fb7b401192c507f7, copied
directly from gpdf. The purpose of this code seems to have been
to provide aliases to icons (so they could be called from 2 different names),
and serve a similar purpose as GtkIconTheme does today.

The whole concept of GtkIconFactory, GtkIconSet, and GtkIconSource was
deprecated in GTK 3.10 back in 2013, so simply get rid of it. This has
no impact in Evince, since the icons can still be found through GtkIconTheme.

To make sure that this is a non-breaking change in libevview, rename
stock_filters-invert icon. The said icon is exported in the API through

and the internal name "stock_filters-invert" was never documented. Therefore,
renaming the icon to "inverted" as exposed by the API should still make the
icon available to consumers (via EV_STOCK_INVERTED_COLORS), without the need
to use GtkIconFactory.