Do not implement module loading through GTypeModule
Instead of using a custom subclass of GTypeModule, load backends directly
using GModule and make the modules always resident. It requires modules to
implement the one-liner ev_backend_query_type so that Evince can query the
type that backends currently register. In the process, completely get rid of
ev_module and port the backends to use default GLib macros.