libview: do not keep track of annotation window position with respect to view
The support for doing this kind of work has been removed in GTK4. The Migration
guide states with respect of functions like "gtk_window_move":

Most likely, you should just stop using them. In some cases, you can fall back
to using the underlying GdkToplevel APIs (for example,
gdk_toplevel_begin_resize()); alternatively, you will need to get the native
windowing system surface from the GtkWindow and call platform specific API.

In addition, the whole logic of having the annotation window be placed in a
fixed position with respect to the main view never worked in Wayland. If this
is a desired feature, it is most likely better to re-implement it from
scratch following the recommendations from GTK, and having Wayland in mind.

Co-developed-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>