Move pps-debug from libdocument to libview
These files, and the debugging and profiling API that they contain are private,
and can only be used while developing Papers. Therefore, it makes sense to place
them where they are used. Since they're only used in libview, but not in
libdocument, move them there. This is a small conceptual simplification that
could help future new developers.