libview: EvJobFind: use EvFindOptions instead of old case_sensitive
The parameter was not deprecated, but might have as well been, as the
comments in the code show. use EvFindOptions as default instead, which
is less error-prone (e.g: cannot provide case_sensitive=TRUE and then
override it), for which support already exists, and which is the one
used internally.