libview: remove fullscreen property in document-model
It's unused in evince, and whether the document requests to be
fullscreen is already available in the EvDocumentMode enum from
EvDocumentInfo. Previously, it was used to sync the fullscreen state of
the window and some of its related widgets. If that were needed in the
future, it would probably make sense to hook the logic directly there,
instead of using ev-document-model as an auxiliary central class for
something that is UI related.