shell: check icon_view exists upon usage in EvSidebarThumbnails
Commit 5278ec22730c7e470b92a97b692f499cb16ee0d3 did a very subtle
change in the way the sidebar is set visible. Before that commit, the
state of the sidebar would we saved in EvWindow chrome property, but
only set to visible once the document was available. After that
commit, the sidebar can be set visible any time, including when the
document is not yet set. In consequence, if the default settings set
the sidebar to visible, EvSidebarThumbnails will try to allocate
before the icon-view is created on the document update. This throws
a critical, that can just be fixed by making sure that we skip the
changes related to the allocation, if the icon-view does not yet
exist.

Most likely, the most elegant change would be to make sure that the
icon-view is created on the widget init. However, for that the model
is also required on init. This requires further refactoring, as all
models should be avaialable otherwise for all the sidebar pages upon
init. This is indeed part of the GTK4 port, so it should probably be
referred to then.

Fixes #1996