libview: rename EvJobRender to EvJobRenderCairo and deprecate it
For the porting to GTK4, it is necessary to have a renderer based on
GdkTexture. However, it was requested by Christian Perch to not
completely remove the cairo renderer, and instead split it in two:
EvJobRenderCairo and EvJobRenderTexture. So do the first part of it
ahead, but deprecating the cairo renderer, that otherwise is not
necessary in evince, and eventually should be ported out for GTK4
applications.

In the process, removed an unused EvJobRender pointer in EvView