libview: rename EvJobThumbnail to EvJobThumbnailCairo and deprecate it
For the porting to GTK4, it is necessary to have a thumbnail based on
a GdkTexture. However, it is also helpful to retain the older cairo
renderer to do the transition at different paces. So instead of
replacing the job with a one based on a textures, split it in two:
EvJobThumbnailCairo and EvJobThumbnailTexture. We do the first part of
it ahead now, but deprecating the cairo thumbnailer, that will
eventually be phased out, as all the depending could should eventually
be ported to the GTK4 native way of working.