shell: derive EvRecentView from GtkBin to simplify GTK4 move
The recent view commit in the GTK4 port is quite big because the
whole machinery behing GtkIconView is deprecated and the API
hard to use. Using directly a GtkGridView and GListStore simplifies
things and is the right move.

However, such commit also includes a reparenting from GtkScrolledWindow
to AdwBin, since the scrolled window is no longer derivable in GTK4. We
can do that work ahead, and simplify the port further.

Co-developed-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>