libmisc: re-parent EvPageActionWidget to a GtkBox
Since bb81f9597a50095d6c20d145d5d460d636d7c8ee evince's toolbar is not
a GtkToolbar, but a GtkBox. Therefore, there's no longer a reason to
have EvPageActionWidget be a GtkToolItem, and can instead be a GtkBox.
This helps the GTK4 port, since GtkToolbar and GtkToolItem have been
removed. This also simplifies a bit the code, as the custom width
check is no longer needed, and the additional box to hold elements is
neither.

In the process, do some small cleanups, like unnecessary function
prototypes at the beginning of the file, or setting properties using
their setters and not g_object_set.

Co-developed-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>