shell: Remove ev_toolbar_get_mode
This is a series clean up aim to remove EvToolbar. EvToolbarMode is
basically identical to EvWindowRunMode and EvWindowPageMode.

The only user of ev_toolbar_get_mode is ev_window_is_recent_view so we
simply extend EvWindowRunMode and query ->window_mode to identify
whether EvWindow is in recent view mode.

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>