shell: Move EvRencentView as stack view child of EvWindow
To make it happend, we remove the open button from EvToolbar and
add an AdwHeaderBar to EvRecentView. The title handling for recent view
is fairly simple so we simply move it to EvRecentView from
EvWindowTitle.

The password view may be broken due to this transition. And we should
fix that later in the transition of EvPasswordView.

In addition of the transition, we changed the style of open button with
more modern AdwButtonContent helper.

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>