shell: Move EvPasswordView to child of view stack of EvWindow
Same as EvRecentView transition except EvWindowPageMode is involved. The
method to switch EvView and EvPasswordView is changed so we don't need
to reference them manually. Let the widget builder keep the references
and release them.

After the transition of EvPasswordView and EvRecentView to stack view,
there is only one NORMAL mode remaining. So we can simply assume
EvToolbar only has one mode which is NORMAL mode.

Same as EvToolbar change, there is only one type of title type left. So
let's remove it.

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>