shell: Switch to the new GtkFileLauncher API
And remove the only usage of libdazzle so we can remove it in the
following commit. Note that current implementation is broken on MacOS
since the meson port and we decide to not consider win32 support.

The new API is an abstract API so we may fix that in GTK side.

Signed-off-by: Qiu Wenbo <qiuwenbo@kylinos.com.cn>