find-sidebar: display spinner during ongoing search job
The find sidebar looks empty during search. It should
signal to the user that the search is ongoing. This makes
the use of the progress bar unnecessary, and thus better
aligns with the HIG, as the order of magnitude in which
a search job is typically finished is too short to
warrant the use of a progress state conveying widget.

This makes the find sidebar a GtkStack that contains
either its results, said spinner, a no results view, or
an initial/cleared state.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>