window: toggle search on Ctrl+F with no text selection
Usually, Ctrl+F toggles the search interface in an app,
in Papers, however, it is a bit more complicated, because
it supports to use it to search for the current selection.

Hence, here, the toggling behaviour is bound to the case
where there is no text selection, while if there is one,
the search bar stays visible and instead is updated with
a new search query.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>