shell: Load previous state of sidebar when a document is opened.
The sidebar is expanded every time a new document is opened. It doesn't remember it's previous state that was set during the last session. Other settings like zoom and window-width are carried over from the previous session. This is a bug because Evince remembers the state of the sidebar, so this needs to be fixed.