Use saner default window sizes for modern displays
When the user document size preferences are not set,
use a more generous default width and height, avoiding tiny contents,
to avoid making the application unhelpful by default.
We should not default to 600x600, this is not the nineties anymore.
If it exceeds the screen's resolution, the WM will clamp it just fine.

This change does not create a smarter sizing heuristic,
it just makes the symptoms less terrible.

Works well with letters (A4 & 8.5x11) at these screen resolutions:

* 2560x1080
* 1920x1080
* 1600x900 and 900x1600
* 1368x768 and 768x1368

In all these cases it behaved exactly as intended, a marked improvement.

The only case where the sizing can arguably be suboptimal is 1080x1920,
but it still behaves better than the previous default.

Even landscape-oriented documents benefit from this bigger sizing.

Related to issue #20.