shell: remove methods in password-view in favor of properties
In preparation for a migration to Rust, we get rid of all methods
from pps-password-view, and interact with it purely through
properties and signals. This is to later avoid having to expose
Rust methods to C code.

The "ask-password" property feels particularly hacky, but it is
just a temporary workaround until pps-window itself is migrated
to Rust.