shell: reorganize actions for PpsWindow
Before this commit, all the PpsWindow actions had the "win"
prefix. However, not all of them had global shortcuts in
PpsApplication. It turns out, many of these actions do not correspond
to the window, but to individual documents in the document view. For
now, these are actually the same. There is one document per Papers
process. In out work towards fixing that, we need to make sure that
the actions belong to each document. This should also simplify the
split between PpsWindow and PpsDocumentView that we are working
towards.