shell: remove user-reload action
The reload feature exists since the beginning of the code-base, but it
probably serves very little purpose as an user action. We currenly
have a monitor that reloads if the file changed outside our
process, which is sensible. The user reload action seems to have been
introduced in 95c733b99f7f8ba9dceabec6085f819cf840ee72 as some import
from nautilus. There it probably makes sense, but it's not like
reloading PDFs is going to change anything (compared to when things
change in your filesystem).