shell: prompt users or reload if document changed outside Papers
Instead of having a custom gsetting for it, reload the document
unconditionally if it's save to do so (no data will be lost). And let
the use choose what to do if the document has changes.

The text for PPS_WINDOW_ACTION_RELOAD can be changed, because since
last commit nothing else uses PPS_WINDOW_ACTION_RELOAD

Beware that the type of the function is also changed, as was wrong
before (connects to signal_swapped)