libview: Check END state when we set PpsViewPresentation to NORMAL
The callback handler of key press event will simply call set_normal if
the key pressed is not specially handled. When the state is already END,
this will make Key_Left trigger a set_normal (in the key pressed
callback) and then call previous_page (in the action handler). The
previous_page function won't work as expected since the state is
inconsistent.

We rename set_normal to set_normal_or_end to reflect the behavior
change.

Fixes #123

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>