shell: search-box: remove previous and next buttons
The action of those buttons are reachable from a keyboard
using the up and down keys on the sidebar, or specific
hotkeys from anywhere in the window.

And, you can click/tap on the surrounding results as
quickly as using the buttons.

Thus, they are not needed, here, but are blocking from
making the search box smaller without reducing the entry's
width rediculously.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>