shell: Pass the attachments list through a GListModel
GList is hard to use as a signal parameter in Rust. The rust side of
code simple drop the reference when the created GListStore goes out of
scope after emitting the popup signal. So we need to reference it in the
C side. Also, GListStore automatically takes a reference of its items,
and release them on free.

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>