libview: ignore adding annotations on move, and markup on release
Since the previous commit, markup (highlight) annotations can only be
added when a text is already selected, so there's no need for these
complete code-paths. For the moving side, it is conditional to
adding_annot_info.annot existing, which is never the case in TEXT
annotations, and not anymore in markup annotations. In addition, there
is no need to take care of markup annotations on release, since they
are handled through the selection code-path that does not need this.

For the same reason, we can also remove the logic on
cancel_add_annotation that needs adding_annot_info.annot, and ignore
the cursor change for MARKUP annotations