libview: rework selection and annotation move with GtkGestureDrag
The logic is split from the callbacks of GtkGestureClick press and
release + GtkEventControllerMotion motion into its own
GtkGestureDrag (which is actually the gesture that takes care of
that). This allows to put a lot more logic local to those callbacks,
making several of the fields in the privates unnecessary. One big
advantage of this is that the global "pressed_button" can be removed
too, and it makes it a lot clearer what do each of the controllers
work for.

TODO:
 - Fix SHIFT selection, take from previous commit

Signed-off-by: Pablo Correa Gomez <ablocorrea@hotmail.com>

SHIFT selection will be fixed in a subsequent commit.

Double tap drag is a common gesture on touchpads, and as
such this needs to stay and its value has to be updated on
every press event (not just n_press > 1, because it needs
to reset), so I reinstate that part of the clean-up.

Co-authored-by: Markus Göllnitz <camelcasenick@bewares.it>
Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>