libview: remove some complexity in PpsView handle_cursor_over_link
Since the previous commit, motion events do not get triggered if there
is a drag event (as they are only relevant as the cursor moves
through the widget). So we don't need this anymore.

Signed-off-by: Pablo Correa Gomez <ablocorrea@hotmail.com>

Remove TODO regarding SHIFT + click selection extension
from click gesture press event handler, because that is
no handled in the separated selection drag gestures from
my commit this one was rebased onto.

Co-authored-by: Markus Göllnitz <camelcasenick@bewares.it>
Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>