shell: window: toggle find only once for CTRL+F with selection
While the state needs to be kept in sync, and activating
the action with a specific state is not possible from
the API for the action itself, it is annoying that the
state is first set to FALSE, to ensure it is toggled back
to TRUE subsequently.

This can lead also to brief flashing of the find sidebar.

I just didn't think ok g_action_group_change_action_state,
but that does exactly what is needed here.

Fixes: 003f229efe52408d35788e9e3006b801d9331d5b ("shell: window: set correct toggle find state")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>