shell: use CSS variables instead of GTK named colours
With libadwaita 1.6, CSS variables are used for colour
variables, making it possible to override them in CSS.
As such, this clean-up provides the practical improvment
that it will pick up accent colour overrides.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>

This needs https://gitlab.gnome.org/GNOME/gtk/-/merge_requests/7306
that is not yet in a release

Signed-off-by: Pablo Correa Gomez <ablocorrea@hotmail.com>

It also now replaces alpha with a single colour mix-in
with transparent.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>