libview: don't sync annotation contents on dispose but on close-request
The annotations windows are not disposed until the main window is
closed, and at that point the document is not saved, so it makes no
sense to sync the annotation results. Saving on closing the annotation
window (which due to hide-on-close is indeed just hidden) makes a lot
more sense.

Fixes #136