shell: remove DocumentLoaded signal from Window dbus interface
This was introduced in a3e68aa381fd1566049a50b26b59ce8e8612f9f7
referencing bugzilla bug #626561. The rationale there was to introduce
it to avoid some problems with FindDocument dbus method, that we
removed in 1dfa9fdd5090849eeda65337f4d733d72e28b2a2 when we got rid of
the daemon.