shell: Fix NULL job returned by PpsSearchContext::finished signal
This is spotted by the Rust binding code of PpsSearchContext. Even
priv->job is checked to not be NULL before entering the branch, we still
always get NULL in the signal handlers. We should fire the signal before
calling clear_job so that priv->job is not set to NULL before passing it
into the signal handlers.

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>