shell: use PpsWindow helper to set PpsDocument of PpsDocumentModel
The previous code would set the document on PpsDocumentModel, and then
have a notify::document signal for doing the setup of the window. The
document itselt cannot really change withouth the window being aware
of it. So there is no need to add such complexity that is hard to
follow, when we can just do all the work synchronously!