libview: remove image drag and drop code
The feature to drag and drop an image was removed in a
previous commit already, except some code in the gesture
click handlers. It would be great to bring back this
feature, and it would not be as entangled as before with
the new drag gestures and gesture claiming in place.

But it should never live in the click gesture, should not
need to keep track of the start coordinates, or whether
the gesture is active, because that is covered by drag
gestures. So, it would only need to have the dragged
image. And that is what is kept in the header file with
everything else here removed.

Link: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/135
See: 39227067140f2e4d02b7b2e62fbf5673cd87dc15 ("libview: drop the ability to drag images and text out of documents")

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>