libview: fix measurement of PpsView
The minimum width of PpsView should always 1 if it should
follow the allocated width of its parent. This always the
case unless it has a specified zoom level (i.e. sizing
mode is PPS_SIZING_FREE).

If the height is also expected to follow the allocated
height of the parent widget (which is only the case with
sizing mode PPS_SIZING_FIT_PAGE), that is to be 1 as well.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/157

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>