libview: context: introduce attachment context
The attachments were not yet controlled in one place.
This is now the place. In a subsequent commit, PpsWindow
and PpsSidebarAttachments will make use of it, instead
of duplicating the work.

The new attachment context dropped storing to non-native
locations as being unnecessary, and only adding to the
complexity.

Note: Not every remote FS is non-native,
e.g. Nextcloud instances are still native, and remain
supported.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>