shell: port to new attachment context
From here on, the state of attachment-aware widgets, are
supposed to be only stored in the attachment context.

For now, one exception to this is tracking selected
attachments, for when their context menu is opened. The
reason for this is that context menus will need to be
refactored later on.

As the new attachment context did not implement storing
to non-native file locations, this is a regression. It
is to reduce complexity and non-native locations are
deemed unnecessary.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>