libdocument: Fix selection region of poppler backend
We need to specify a list of quadrilaterals to create a text markup
annotation. Our implementation of computing this list use the wrong
interface to compute the list through the rectangles of cairo region.
Although this working as expected in most time, the rectangles can be
merged to form a larger rectangle which make the quadrilaterals not
precisely reflect the text area.

We fix this issue by using a reference implementation from poppler demo.

Closes #184

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>