shell: hide search button when search not available
The search button is not useful without a search feature,
so it makes sense to just hide the button instead of
making the user aware of a feature not available.

If someone was looking for the search button there, from
its absence it should be easily deducible the feature is
available.

The find unsupported icon is now not used anymore, and
thus removed.

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>