libdocument: Handle annotation with different char size
It's common that one line of document has different size of character.
Our old behavior is treating this line as a single quadrilateral. But
with !265 we regard them as several different quadrilaterals because
they don't have a common height. With this commit, we can regard them as
a single line when certain condition hold.

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>