shell: Move the Esc handler from the application to the widget
We should not register Escape as a application level global shortcut
since we use AdwDialog. So we can move the shortcut into the PpsWindow
widget. The find sidebar uses GtkSearchEntry which register a shortcut
of Escape to trigger the stop-search signal, so we listen to the signal
and invoke the doc.find-toggle action. We don't need to check the
focused child of find sidebar since the Escape key is already handled by
the GtkSearchEntry when it's focused.

Closes #170
Closes #193
Closes #194

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>