shell: Refactor PpsWindow to PpsDocumentView
By:
  1. Rename PpsWindow to PpsDocumentView
  2. Remove other views from PpsDocumentView
  3. Remove GtkWindow related stuff from PpsDocumentView

This commit simply remove all views from PpsDocumentView. So baiscally
we won't let it manage all the views anymore. As the widget name,
PpsDocumentView will only use to show the normal view we called before.

This commit also remove the load_job and reload_job. Which means we will
use the true PpsWindow to mange the file load, reload and file monitor.

This widget is no longer a GtkWindow so some of its implementation
should be removed.

Signed-off-by: Qiu Wenbo <qiuwenbo@gnome.org>