shell: use settings instead of metadat for setting the sidebar page
The document view should hold itself the status of the sidebar. In the
case where we would have multiple documents, it would be unexpected if
changing the document would change the visible sidebar. Therefore,
store the state in the settings and not in the metadata. We do so by
binding the visible-child-name property and the setting, and thus we
can avoid having to explicitly save the setting on close.