shell: use settings instead of metadata to make sidebar visible
The document view should hold itself the status of the sidebar. In the
case where we would have multiple documents, it would be unexpected if
changing the document would change the visibility of the
sidebar. Therefore, store the state in the settings and not in the
metadata. Contrary to the sidebar-page, we cannot directly bind the
visibility, since the "collapsed" state adds more complexity to it.