shell: document-view: move focus to view only when sidebar is collapsed
When a user navigates the view from inside the sidebar,
the sidebar will be collapsed when it is overlaid, and
then it makes sense to focus the view.

However, when the sidebar is next to the view,
interacting with it, is – conceptually – not something
transient, so the focus should remain in the sidebar.

Additionally, when the view is not implicitly interacted
with – which is not always the case – when the sidebar
is closed, there is no reason to prefer focussing the
view over any other UI component.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/219

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>