shell: document-view: remove unnecessary signal action update
The signal handler should have been connected as
G_CONNECT_SWAPPED, as otherwise the signature mismatch
leads to a crash.

However, the signal handling can be removed altogether,
as the document actions do not depend on the search being
active or not – besides the find-prev/-next, that is, but
they are not handled in
pps_document_view_update_actions_sensitivity.

Fixes: e686a67c1ae8a6210951bd0888093a1d0293b177 ("libview+shell: move view search integration from window into view")
Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/229

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>