Drop Windows support
Fixes https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/3

As documented in that issue, the whole thing doesn't build, and
there's no maintainer or contributor interested. We would be happy to
revert this if somebody showed up, but otherwise this code is
unreachable, and has been for too long. So drop it.x