shell-rs: sidebar-links: do not all deep-expanding leaves
Leaf outline entries do not have children, and their
child list model does not exist (and is not just an empty
one). Thus, trying to deep-expand a leaf entry makes
Papers crash at unwrapping the child model.

Conceptually, it does not make sense to expand something
that is not possible to be expanded. Thus, the action
should be disabled.

Closes: https://gitlab.gnome.org/GNOME/Incubator/papers/-/issues/251

Signed-off-by: Markus Göllnitz <camelcasenick@bewares.it>