shell: replace loading overlay with showing LoaderView during load
This allows to simplify the document view, and provides a more
standard UX for this use-case, where an Overlay is not common.
We do so reusing the PpsLoaderView. If we have a positive number, then
use a progress bar, else use an AdwSpinner. This would allow to, in
the future, add a callback that sets the progress for local document
types.

This possibly regress on showing the overlay when single pages that
take long to render are loading. However, this never worked out well,
since the overlay was not on the page, but on the document view. In
the future, when we have each page be a widget, we could very easily
fix that with a spinner on the page