libview: remove PpsView is-loading property and all code associated
The view is not the widget that should take care of notifying if
something is or isn't loading. What keeps track of that is the model
and the jobs (for which the window is ultimately responsible). There
is no reason for this information to be propagated up, and we stopped
using this in the previous commit. This was
initially introduced in 035c1cbd4756d3819a6616f62acd87662abc6374
with the reasoning to move the implementation of the Loading widget to
the shell (it still is there!), and allowing users to implement their
own loading notification systems. This is certainly not what we
want. For documents taking long to load, since some commits back, we
have now a spinner on the window to take care of it.

In the process, also remove some private pointers that had been unused
since the is-loading property was introduced