libview: simplify dealing with adding unsupported attachment annots
all the other annotation types are markup annotations, so we can get
rid of an unnecessary if. This is reasonable given that nobody has
looked at how to add attachment annotations in 15 years. We might as
well just stop assuming we will add support soon, and if somebody uses
it, then they'll bump into an assert_not_reached