libview: remove SIGNAL_ANNOT_ADDED and further integration logic in shell
Since the previous commits, we now add annotations in a single action,
instead of starting it, and then letting the view signal when it's
done. This requires an extra missing check to make sure we notify
the document view if the annotation failed to create. In general, this
reduces the integration and the coupling.

Helps #137