libdocument: backends: remove XPS support
This document format has seen no maintenance in the last 10 years, and
virtually no changes in the last 14. This is a document format that is
really rare in 2024 (most people won't have seen one), and it should
be possible to use different tools to convert these files to
PDF. There is no strong justification for maintaining this code
in-tree. Would be fine to help others maintain it out-of-tree, though