shell: do not update actions sensitivity at init
Since we moved the actions to a doc namespace, this has become
unnecessary. This function is run again when we set the document, and
document actions are triggered by either buttons that are only
available in the document view, or by shortcuts that now are only
enabled if the focus is on it. Therefore, this call simply does
nothing.