shell: keep page-related actions always enabled
At some point, these actions were all enabled and disabled to
conditionally show navigation actions in the main popup menu. The last
actions are, since previous commit, removed. And the shortcuts
attached to them are only in action if there is a document, so there's
no risk of these being enabled without a document