libdocument: Make PpsCertificateInfo a class
Make PpsCertificateInfo a class as we are adding quite a lot of
members to the class and handling them as properties is
more convenient in Rust.

Also remove some getters as these usually return "const"
values but when getting a property via g_object_get() these
are duplicated/reffed so once this get into Rust it would be
inconsistent as Rust could implement getting of the property
via the getter or via the g_object_get(). This also shrinks API
a little.