signatures: Move signature details to a tab in properties
This commit moves signature details to a tab in properties. It originates
from my work on showing of signatures made for Evince.
It implements UI handling in Rust and tries to be as close to mock-ups created
for Evince as possible (see https://gitlab.gnome.org/Teams/Design/app-mockups/-/issues/32#note_1447378).
Compared to those mock-ups it does not work with revisions but with signatures.