libview: use the search context to get results to highlight
Instead of hooking up to the job itself. This reduces the complexity
remarkable and reduces some risks like sending the job outside of the
context scope. This means that now the search results are only
highlighted as the model gets updated. For now, this means at the end
of the search, but this would also work if the model gets filled on
updated callbacks like proposed in
https://gitlab.gnome.org/GNOME/Incubator/papers/-/merge_requests/389